Отрисовка компонентов страницы разбита на отдельные модули
