+++ src/modules/render-user-rank.js
@@ -1,0 +1,8 @@
+import {getUserRankHtml} from '../view/user-rank';
+import {insertHtmlElement} from '../services/utils';
+
+export const renderUserRank = () => {
+  const userRankContainer = document.querySelector('.header');
+  const userRank = getUserRankHtml();
+  insertHtmlElement(userRankContainer, userRank, 'beforeend');
+};
